django/pyproject.toml
[tool.poetry]
name = "danakka"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
pymysql = "^1.0.2"
cryptography = "^39.0.0"
pillow = "^9.4.0"
requests = "^2.28.2"
korean-lunar-calendar = "^0.3.1"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
